# docs/swagger.yaml
info:
  contact: {}
paths:
  /index:
    get:
      description: 首页
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: 首页
      tags:
      - 首页信息
  /user/createUser:
    get:
      description: 创建用户
      parameters:
      - description: 用户名
        in: query
        name: name
        type: string
      - description: 密码
        in: query
        name: password
        type: string
      - description: 确认密码
        in: query
        name: repassword
        type: string
      - description: 手机号
        in: query
        name: phone
        type: string
      - description: 邮箱
        in: query
        name: email
        type: string
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: 创建用户
      tags:
      - 用户模块
  /user/deleteUser:
    get:
      description: 删除用户
      parameters:
      - description: 用户名
        in: query
        name: name
        type: string
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: 删除用户
      tags:
      - 用户模块
  /user/getUserList:
    get:
      description: 获取用户信息
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: 获取用户信息
      tags:
      - 用户模块
  /user/updateUser:
    post:
      description: 修改用户
      parameters:
      - description: id
        in: formData
        name: id
        type: string
      - description: 用户名
        in: formData
        name: name
        type: string
      - description: 密码
        in: formData
        name: password
        type: string
      - description: 手机号
        in: formData
        name: phone
        type: string
      - description: 邮箱
        in: formData
        name: email
        type: string
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: 修改用户
      tags:
      - 用户模块
  /user/userLogin:
    post:
      description: 用户登入
      parameters:
      - description: 用户名
        in: formData
        name: name
        type: string
      - description: 密码
        in: formData
        name: password
        type: string
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: 用户登入
      tags:
      - 用户模块
swagger: "2.0"